Add bcrypt for password hashing; refactor user model and controller for improved file handling and validation; update dotenv configuration paths

+++ server/src/config/cloudinary.js
@@ -1,7 +1,12 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
 
-// Configuration
+dotenv.config({
+    path: "./.env",
+});
+
+
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
@@ -12,24 +17,33 @@ const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null;
 
-        //upload the file on cloudinary
+        
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto",
         });
-        // file has been uploaded successsfully
-        // console.log("file is uploaded on cloudinary", response.url);
+
+        
         fs.unlinkSync(localFilePath);
         return response;
+
     } catch (error) {
-        fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+       
+        console.error("CLOUDINARY UPLOAD ERROR:", error); 
+
+        
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        
         return null;
     }
 };
 
 const deleteFromCloudinary = async (publicId, resourceType = "image") => {
     try {
-        // Delete the asset with the given public ID
-        const response = await cloudinary.uploader.destroy(publicId, {resource_type: resourceType});
+        const response = await cloudinary.uploader.destroy(publicId, {
+            resource_type: resourceType
+        });
         return response;
     } catch (error) {
         console.error("Error deleting asset from Cloudinary:", error.message);
